backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
kiteconnect==3.10.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
aiohttp==3.9.1
